+++ routes/web.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Route;
 use Simtabi\Laranail\Ichava\Browser\Http\Controllers\Web\IconBrowserController;
-use Simtabi\Laranail\Ichava\Support\Helpers;
 
 /*
 |--------------------------------------------------------------------------